Added support for slashes within quoted strings
The parser jumped to the tagEndHandler once it received a '/' within a
quoted string.

A fourth test case was added to cover an example gpx-type file, which
has attributes with URL strings, containing forward slashes.

Another issue was that the tests never failed, due to some issues in
the cmake configuration. The test program also did not return a
non-zero on failure, letting the tests falsely succeed, even if they
failed.